requirements.txt
fastapi
uvicorn
pdfplumber
chromadb
google-generativeai
python-multipart
langchain
langchain.text_splitters
